refactor: cli and controller
- move CliRunner and CliFlags to folder cli, update imports and tests to match
- move PreprocessorController and Result to folder controller.
- remove Object from Result constructor.
